requirements.txt
# Core Dependencies
openai>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
asyncio
concurrent.futures

# XML Processing
lxml>=4.9.0
beautifulsoup4>=4.12.0
xmltodict>=0.13.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
networkx>=3.1.0

# Graph Database
neo4j>=5.13.0
py2neo>=2021.2.3

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Monitoring and Logging
loguru>=0.7.0
prometheus-client>=0.18.0
rich>=13.6.0

# Configuration
pydantic>=2.4.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Utilities
tqdm>=4.66.0
psutil>=5.9.0
schedule>=1.2.0

# Development
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0 